Rename all mat2 materials → mat_
