+++ Animate/materials/mat_forward_colored_gouraud_skinned.ts
@@ -1,48 +1,41 @@
-import {link, Material} from "../common/material.js";
-import {GL_TRIANGLES} from "../common/webgl.js";
-import {ColoredShadedLayout, ForwardShadingLayout} from "./layout.js";
+import {link, Material} from "../../common/material.js";
+import {GL_TRIANGLES} from "../../common/webgl.js";
+import {ColoredShadedLayout, ForwardShadingLayout} from "../../materials/layout.js";
+import {SkinningLayout} from "./layout_skinning.js";
 
 let vertex = `#version 300 es\n
-
-    uniform mat4 pv;
-    uniform mat4 world;
-    uniform mat4 self;
-
-    in vec3 attr_position;
-    in vec3 attr_normal;
-
-    out vec4 vert_position;
-    out vec3 vert_normal;
-
-    void main() {
-        vert_position = world * vec4(attr_position, 1.0);
-        vert_normal = (vec4(attr_normal, 1.0) * self).xyz;
-        gl_Position = pv * vert_position;
-    }
-`;
-
-let fragment = `#version 300 es\n
-    precision mediump float;
 
     // See Game.LightPositions and Game.LightDetails.
     const int MAX_LIGHTS = 8;
 
+    uniform mat4 pv;
+    uniform mat4 world;
+    uniform mat4 self;
     uniform vec3 eye;
     uniform vec4 diffuse_color;
     uniform vec4 specular_color;
     uniform float shininess;
     uniform vec4 light_positions[MAX_LIGHTS];
     uniform vec4 light_details[MAX_LIGHTS];
+    uniform mat4 bones[6];
 
-    in vec4 vert_position;
-    in vec3 vert_normal;
+    in vec3 attr_position;
+    in vec3 attr_normal;
+    in vec4 attr_weights;
 
-    out vec4 frag_color;
+    out vec4 vert_color;
+
+    mat4 world_weighted(vec4 weights) {
+        return weights[1] * bones[int(weights[0])] + weights[3] * bones[int(weights[2])];
+    }
 
     void main() {
-        vec3 world_normal = normalize(vert_normal);
+        mat4 bone_world = world_weighted(attr_weights);
+        vec4 world_position = bone_world * vec4(attr_position, 1.0);
+        vec3 world_normal = normalize(mat3(bone_world) * attr_normal);
+        gl_Position = pv * world_position;
 
-        vec3 view_dir = eye - vert_position.xyz;
+        vec3 view_dir = eye - world_position.xyz;
         vec3 view_normal = normalize(view_dir);
 
         // Ambient light.
@@ -61,7 +54,7 @@ let fragment = `#version 300 es\n
                 // Directional light.
                 light_normal = light_positions[i].xyz;
             } else {
-                vec3 light_dir = light_positions[i].xyz - vert_position.xyz;
+                vec3 light_dir = light_positions[i].xyz - world_position.xyz;
                 float light_dist = length(light_dir);
                 light_normal = light_dir / light_dist;
                 // Distance attenuation.
@@ -74,11 +67,6 @@ let fragment = `#version 300 es\n
                 light_acc += diffuse_color.rgb * diffuse_factor * light_color * light_intensity;
 
                 if (shininess > 0.0) {
-                    // Phong reflection model.
-                    // vec3 r = reflect(-light_normal, world_normal);
-                    // float specular_angle = max(dot(r, view_normal), 0.0);
-                    // float specular_factor = pow(specular_angle, shininess);
-
                     // Blinn-Phong reflection model.
                     vec3 h = normalize(light_normal + view_normal);
                     float specular_angle = max(dot(h, world_normal), 0.0);
@@ -90,13 +78,25 @@ let fragment = `#version 300 es\n
             }
         }
 
-        frag_color = vec4(light_acc, 1.0);
+        vert_color = vec4(light_acc, 1.0);
     }
 `;
 
-export function mat2_forward_colored_phong(
+let fragment = `#version 300 es\n
+    precision mediump float;
+
+    in vec4 vert_color;
+
+    out vec4 frag_color;
+
+    void main() {
+        frag_color = vert_color;
+    }
+`;
+
+export function mat_forward_colored_gouraud_skinned(
     gl: WebGL2RenderingContext
-): Material<ColoredShadedLayout & ForwardShadingLayout> {
+): Material<ColoredShadedLayout & ForwardShadingLayout & SkinningLayout> {
     let program = link(gl, vertex, fragment);
     return {
         Mode: GL_TRIANGLES,
@@ -116,6 +116,9 @@ export function mat2_forward_colored_phong(
 
             VertexPosition: gl.getAttribLocation(program, "attr_position")!,
             VertexNormal: gl.getAttribLocation(program, "attr_normal")!,
+
+            Bones: gl.getUniformLocation(program, "bones")!,
+            VertexWeights: gl.getAttribLocation(program, "attr_weights")!,
         },
     };
 }
