+++ Instancing/materials/mat_forward_instanced.ts
@@ -1,6 +1,6 @@
-import {link, Material} from "../common/material.js";
-import {GL_TRIANGLES} from "../common/webgl.js";
-import {ColoredShadedLayout, ForwardShadingLayout} from "./layout.js";
+import {link, Material} from "../../common/material.js";
+import {GL_TRIANGLES} from "../../common/webgl.js";
+import {ForwardInstancedLayout} from "./layout_instancing.js";
 
 let vertex = `#version 300 es\n
 
@@ -10,27 +10,26 @@ let vertex = `#version 300 es\n
     uniform mat4 pv;
     uniform mat4 world;
     uniform mat4 self;
-    uniform vec3 eye;
-    uniform vec4 diffuse_color;
-    uniform vec4 specular_color;
-    uniform float shininess;
+    uniform vec3 palette[16];
+
+    uniform int light_count;
     uniform vec4 light_positions[MAX_LIGHTS];
     uniform vec4 light_details[MAX_LIGHTS];
 
     in vec3 attr_position;
     in vec3 attr_normal;
+    in vec4 attr_offset;
+
     out vec4 vert_color;
 
     void main() {
-        vec4 attr_pos = world * vec4(attr_position, 1.0);
-        vec3 attr_normal = normalize((vec4(attr_normal, 1.0) * self).xyz);
-        gl_Position = pv * attr_pos;
-
-        vec3 view_dir = eye - attr_pos.xyz;
-        vec3 view_normal = normalize(view_dir);
+        vec4 world_position = world * vec4(attr_position + attr_offset.xyz, 1.0);
+        vec3 world_normal = normalize((vec4(attr_normal, 0.0) * self).xyz);
+        gl_Position = pv * world_position;
 
         // Ambient light.
-        vec3 light_acc = diffuse_color.rgb * 0.1;
+        vec3 color = palette[int(attr_offset[3])];
+        vec3 light_acc = color * 0.1;
 
         for (int i = 0; i < MAX_LIGHTS; i++) {
             if (light_positions[i].w == 0.0) {
@@ -45,27 +44,17 @@ let vertex = `#version 300 es\n
                 // Directional light.
                 light_normal = light_positions[i].xyz;
             } else {
-                vec3 light_dir = light_positions[i].xyz - attr_pos.xyz;
+                vec3 light_dir = light_positions[i].xyz - world_position.xyz;
                 float light_dist = length(light_dir);
                 light_normal = light_dir / light_dist;
                 // Distance attenuation.
                 light_intensity /= (light_dist * light_dist);
             }
 
-            float diffuse_factor = dot(attr_normal, light_normal);
+            float diffuse_factor = dot(world_normal, light_normal);
             if (diffuse_factor > 0.0) {
                 // Diffuse color.
-                light_acc += diffuse_color.rgb * diffuse_factor * light_color * light_intensity;
-
-                if (shininess > 0.0) {
-                    // Blinn-Phong reflection model.
-                    vec3 h = normalize(light_normal + view_normal);
-                    float specular_angle = max(dot(h, attr_normal), 0.0);
-                    float specular_factor = pow(specular_angle, shininess);
-
-                    // Specular color.
-                    light_acc += specular_color.rgb * specular_factor * light_color * light_intensity;
-                }
+                light_acc += color * diffuse_factor * light_color * light_intensity;
             }
         }
 
@@ -74,6 +63,7 @@ let vertex = `#version 300 es\n
 `;
 
 let fragment = `#version 300 es\n
+
     precision mediump float;
 
     in vec4 vert_color;
@@ -85,9 +75,9 @@ let fragment = `#version 300 es\n
     }
 `;
 
-export function mat2_forward_colored_gouraud(
+export function mat_forward_instanced(
     gl: WebGL2RenderingContext
-): Material<ColoredShadedLayout & ForwardShadingLayout> {
+): Material<ForwardInstancedLayout> {
     let program = link(gl, vertex, fragment);
     return {
         Mode: GL_TRIANGLES,
@@ -97,16 +87,14 @@ export function mat2_forward_colored_gouraud(
             World: gl.getUniformLocation(program, "world")!,
             Self: gl.getUniformLocation(program, "self")!,
 
-            DiffuseColor: gl.getUniformLocation(program, "diffuse_color")!,
-            SpecularColor: gl.getUniformLocation(program, "specular_color")!,
-            Shininess: gl.getUniformLocation(program, "shininess")!,
+            Palette: gl.getUniformLocation(program, "palette")!,
 
-            Eye: gl.getUniformLocation(program, "eye")!,
             LightPositions: gl.getUniformLocation(program, "light_positions")!,
             LightDetails: gl.getUniformLocation(program, "light_details")!,
 
             VertexPosition: gl.getAttribLocation(program, "attr_position")!,
             VertexNormal: gl.getAttribLocation(program, "attr_normal")!,
+            VertexOffset: gl.getAttribLocation(program, "attr_offset")!,
         },
     };
 }
