+++ Particles/game.ts
@@ -1,7 +1,7 @@
 import {Game3D} from "../common/game.js";
 import {Entity} from "../common/world.js";
-import {mat2_forward_particles_colored} from "./materials/mat2_forward_particles_colored.js";
-import {mat2_forward_particles_textured} from "./materials/mat2_forward_particles_textured.js";
+import {mat_forward_particles_colored} from "./materials/mat_forward_particles_colored.js";
+import {mat_forward_particles_textured} from "./materials/mat_forward_particles_textured.js";
 import {sys_camera} from "./systems/sys_camera.js";
 import {sys_particles} from "./systems/sys_particles.js";
 import {sys_render_forward} from "./systems/sys_render_forward.js";
@@ -13,8 +13,8 @@ import {World} from "./world.js";
 export class Game extends Game3D {
     World = new World();
 
-    MaterialParticlesColored = mat2_forward_particles_colored(this.Gl);
-    MaterialParticlesTextured = mat2_forward_particles_textured(this.Gl);
+    MaterialParticlesColored = mat_forward_particles_colored(this.Gl);
+    MaterialParticlesTextured = mat_forward_particles_textured(this.Gl);
     Textures: Record<string, WebGLTexture> = {};
 
     Cameras: Array<Entity> = [];
