+++ RenderTexture/game.ts
@@ -1,7 +1,7 @@
 import {create_forward_target} from "../common/framebuffer.js";
 import {Game3D} from "../common/game.js";
 import {Entity} from "../common/world.js";
-import {mat2_forward_textured_unlit} from "../materials/mat2_forward_textured_unlit.js";
+import {mat_forward_textured_unlit} from "../materials/mat_forward_textured_unlit.js";
 import {mesh_cube} from "../meshes/cube.js";
 import {mesh_plane} from "../meshes/plane.js";
 import {sys_camera} from "./systems/sys_camera.js";
@@ -15,7 +15,7 @@ import {World} from "./world.js";
 export class Game extends Game3D {
     World = new World();
 
-    MaterialTexturedUnlit = mat2_forward_textured_unlit(this.Gl);
+    MaterialTexturedUnlit = mat_forward_textured_unlit(this.Gl);
     MeshCube = mesh_cube(this.Gl);
     MeshPlane = mesh_plane(this.Gl);
