+++ materials/mat_forward_colored_gouraud.ts
@@ -1,8 +1,9 @@
 import {link, Material} from "../common/material.js";
 import {GL_TRIANGLES} from "../common/webgl.js";
-import {ForwardShadingLayout, TexturedShadedLayout} from "./layout.js";
+import {ColoredShadedLayout, ForwardShadingLayout} from "./layout.js";
 
 let vertex = `#version 300 es\n
+
     // See Game.LightPositions and Game.LightDetails.
     const int MAX_LIGHTS = 8;
 
@@ -17,18 +18,15 @@ let vertex = `#version 300 es\n
     uniform vec4 light_details[MAX_LIGHTS];
 
     in vec3 attr_position;
-    in vec2 attr_texcoord;
     in vec3 attr_normal;
-
-    out vec2 vert_texcoord;
     out vec4 vert_color;
 
     void main() {
-        vec4 world_position = world * vec4(attr_position, 1.0);
-        vec3 world_normal = normalize((vec4(attr_normal, 1.0) * self).xyz);
-        gl_Position = pv * world_position;
+        vec4 attr_pos = world * vec4(attr_position, 1.0);
+        vec3 attr_normal = normalize((vec4(attr_normal, 1.0) * self).xyz);
+        gl_Position = pv * attr_pos;
 
-        vec3 view_dir = eye - world_position.xyz;
+        vec3 view_dir = eye - attr_pos.xyz;
         vec3 view_normal = normalize(view_dir);
 
         // Ambient light.
@@ -47,14 +45,14 @@ let vertex = `#version 300 es\n
                 // Directional light.
                 light_normal = light_positions[i].xyz;
             } else {
-                vec3 light_dir = light_positions[i].xyz - world_position.xyz;
+                vec3 light_dir = light_positions[i].xyz - attr_pos.xyz;
                 float light_dist = length(light_dir);
                 light_normal = light_dir / light_dist;
                 // Distance attenuation.
                 light_intensity /= (light_dist * light_dist);
             }
 
-            float diffuse_factor = dot(world_normal, light_normal);
+            float diffuse_factor = dot(attr_normal, light_normal);
             if (diffuse_factor > 0.0) {
                 // Diffuse color.
                 light_acc += diffuse_color.rgb * diffuse_factor * light_color * light_intensity;
@@ -62,7 +60,7 @@ let vertex = `#version 300 es\n
                 if (shininess > 0.0) {
                     // Blinn-Phong reflection model.
                     vec3 h = normalize(light_normal + view_normal);
-                    float specular_angle = max(dot(h, world_normal), 0.0);
+                    float specular_angle = max(dot(h, attr_normal), 0.0);
                     float specular_factor = pow(specular_angle, shininess);
 
                     // Specular color.
@@ -72,29 +70,24 @@ let vertex = `#version 300 es\n
         }
 
         vert_color = vec4(light_acc, 1.0);
-        vert_texcoord = attr_texcoord;
     }
 `;
 
 let fragment = `#version 300 es\n
     precision mediump float;
 
-    uniform sampler2D diffuse_map;
-
-    in vec2 vert_texcoord;
     in vec4 vert_color;
 
     out vec4 frag_color;
 
     void main() {
-        vec4 tex_color = texture(diffuse_map, vert_texcoord);
-        frag_color = vert_color * tex_color;
+        frag_color = vert_color;
     }
 `;
 
-export function mat2_forward_textured_gouraud(
+export function mat_forward_colored_gouraud(
     gl: WebGL2RenderingContext
-): Material<TexturedShadedLayout & ForwardShadingLayout> {
+): Material<ColoredShadedLayout & ForwardShadingLayout> {
     let program = link(gl, vertex, fragment);
     return {
         Mode: GL_TRIANGLES,
@@ -104,7 +97,6 @@ export function mat2_forward_textured_gouraud(
             World: gl.getUniformLocation(program, "world")!,
             Self: gl.getUniformLocation(program, "self")!,
 
-            DiffuseMap: gl.getUniformLocation(program, "diffuse_map")!,
             DiffuseColor: gl.getUniformLocation(program, "diffuse_color")!,
             SpecularColor: gl.getUniformLocation(program, "specular_color")!,
             Shininess: gl.getUniformLocation(program, "shininess")!,
@@ -114,7 +106,6 @@ export function mat2_forward_textured_gouraud(
             LightDetails: gl.getUniformLocation(program, "light_details")!,
 
             VertexPosition: gl.getAttribLocation(program, "attr_position")!,
-            VertexTexCoord: gl.getAttribLocation(program, "attr_texcoord")!,
             VertexNormal: gl.getAttribLocation(program, "attr_normal")!,
         },
     };
