+++ materials/mat_forward_colored_unlit.ts
@@ -1,9 +1,8 @@
 import {link, Material} from "../common/material.js";
-import {GL_TRIANGLES} from "../common/webgl.js";
-import {DepthMappingLayout} from "./layout.js";
+import {GL_LINE_LOOP, GL_LINE_STRIP, GL_TRIANGLES} from "../common/webgl.js";
+import {ColoredUnlitLayout} from "./layout.js";
 
 let vertex = `#version 300 es\n
-
     uniform mat4 pv;
     uniform mat4 world;
 
@@ -17,24 +16,38 @@ let vertex = `#version 300 es\n
 let fragment = `#version 300 es\n
     precision mediump float;
 
+    uniform vec4 color;
+
     out vec4 frag_color;
 
     void main() {
-        // Visualization only. Actual z is saved in the depth buffer.
-        float z = gl_FragCoord.z * 10.0;
-        frag_color = vec4(z, z, z, 1.0);
+        frag_color = color;
     }
 `;
 
-export function mat2_forward_depth(gl: WebGL2RenderingContext): Material<DepthMappingLayout> {
+export function mat_forward_colored_unlit(
+    gl: WebGL2RenderingContext,
+    mode: GLenum = GL_TRIANGLES
+): Material<ColoredUnlitLayout> {
     let program = link(gl, vertex, fragment);
     return {
-        Mode: GL_TRIANGLES,
+        Mode: mode,
         Program: program,
         Locations: {
             Pv: gl.getUniformLocation(program, "pv")!,
             World: gl.getUniformLocation(program, "world")!,
+
+            Color: gl.getUniformLocation(program, "color")!,
+
             VertexPosition: gl.getAttribLocation(program, "attr_position")!,
         },
     };
+}
+
+export function mat_forward_colored_wireframe(gl: WebGL2RenderingContext) {
+    return mat_forward_colored_unlit(gl, GL_LINE_LOOP);
+}
+
+export function mat_forward_colored_line(gl: WebGL2RenderingContext) {
+    return mat_forward_colored_unlit(gl, GL_LINE_STRIP);
 }
