+++ materials/mat_forward_depth.ts
@@ -1,6 +1,6 @@
 import {link, Material} from "../common/material.js";
-import {GL_POINTS} from "../common/webgl.js";
-import {ColoredUnlitLayout} from "./layout.js";
+import {GL_TRIANGLES} from "../common/webgl.js";
+import {DepthMappingLayout} from "./layout.js";
 
 let vertex = `#version 300 es\n
 
@@ -11,35 +11,29 @@ let vertex = `#version 300 es\n
 
     void main() {
         gl_Position = pv * world * vec4(attr_position, 1.0);
-        gl_PointSize = 8.0;
     }
 `;
 
 let fragment = `#version 300 es\n
     precision mediump float;
 
-    uniform vec4 color;
-
     out vec4 frag_color;
 
     void main() {
-        frag_color = color;
+        // Visualization only. Actual z is saved in the depth buffer.
+        float z = gl_FragCoord.z * 10.0;
+        frag_color = vec4(z, z, z, 1.0);
     }
 `;
 
-export function mat2_forward_colored_points(
-    gl: WebGL2RenderingContext
-): Material<ColoredUnlitLayout> {
+export function mat_forward_depth(gl: WebGL2RenderingContext): Material<DepthMappingLayout> {
     let program = link(gl, vertex, fragment);
     return {
-        Mode: GL_POINTS,
+        Mode: GL_TRIANGLES,
         Program: program,
         Locations: {
             Pv: gl.getUniformLocation(program, "pv")!,
             World: gl.getUniformLocation(program, "world")!,
-
-            Color: gl.getUniformLocation(program, "color")!,
-
             VertexPosition: gl.getAttribLocation(program, "attr_position")!,
         },
     };
