+++ materials/mat_forward_textured_gouraud.ts
@@ -1,35 +1,38 @@
-import {link, Material} from "../../common/material.js";
-import {GL_TRIANGLES} from "../../common/webgl.js";
-import {ForwardInstancedLayout} from "./layout_instancing.js";
+import {link, Material} from "../common/material.js";
+import {GL_TRIANGLES} from "../common/webgl.js";
+import {ForwardShadingLayout, TexturedShadedLayout} from "./layout.js";
 
 let vertex = `#version 300 es\n
-
     // See Game.LightPositions and Game.LightDetails.
     const int MAX_LIGHTS = 8;
 
     uniform mat4 pv;
     uniform mat4 world;
     uniform mat4 self;
-    uniform vec3 palette[16];
-
-    uniform int light_count;
+    uniform vec3 eye;
+    uniform vec4 diffuse_color;
+    uniform vec4 specular_color;
+    uniform float shininess;
     uniform vec4 light_positions[MAX_LIGHTS];
     uniform vec4 light_details[MAX_LIGHTS];
 
     in vec3 attr_position;
+    in vec2 attr_texcoord;
     in vec3 attr_normal;
-    in vec4 attr_offset;
 
+    out vec2 vert_texcoord;
     out vec4 vert_color;
 
     void main() {
-        vec4 world_position = world * vec4(attr_position + attr_offset.xyz, 1.0);
-        vec3 world_normal = normalize((vec4(attr_normal, 0.0) * self).xyz);
+        vec4 world_position = world * vec4(attr_position, 1.0);
+        vec3 world_normal = normalize((vec4(attr_normal, 1.0) * self).xyz);
         gl_Position = pv * world_position;
 
+        vec3 view_dir = eye - world_position.xyz;
+        vec3 view_normal = normalize(view_dir);
+
         // Ambient light.
-        vec3 color = palette[int(attr_offset[3])];
-        vec3 light_acc = color * 0.1;
+        vec3 light_acc = diffuse_color.rgb * 0.1;
 
         for (int i = 0; i < MAX_LIGHTS; i++) {
             if (light_positions[i].w == 0.0) {
@@ -54,32 +57,44 @@ let vertex = `#version 300 es\n
             float diffuse_factor = dot(world_normal, light_normal);
             if (diffuse_factor > 0.0) {
                 // Diffuse color.
-                light_acc += color * diffuse_factor * light_color * light_intensity;
+                light_acc += diffuse_color.rgb * diffuse_factor * light_color * light_intensity;
+
+                if (shininess > 0.0) {
+                    // Blinn-Phong reflection model.
+                    vec3 h = normalize(light_normal + view_normal);
+                    float specular_angle = max(dot(h, world_normal), 0.0);
+                    float specular_factor = pow(specular_angle, shininess);
+
+                    // Specular color.
+                    light_acc += specular_color.rgb * specular_factor * light_color * light_intensity;
+                }
             }
         }
 
         vert_color = vec4(light_acc, 1.0);
+        vert_texcoord = attr_texcoord;
     }
 `;
 
 let fragment = `#version 300 es\n
-
     precision mediump float;
 
+    uniform sampler2D diffuse_map;
+
+    in vec2 vert_texcoord;
     in vec4 vert_color;
 
     out vec4 frag_color;
 
     void main() {
-        frag_color = vert_color;
+        vec4 tex_color = texture(diffuse_map, vert_texcoord);
+        frag_color = vert_color * tex_color;
     }
 `;
 
-// Instanced drawing can also be used in WebGL1 via an extension:
-// https://developer.mozilla.org/en-US/docs/Web/API/ANGLE_instanced_arrays
-export function mat2_forward_instanced(
+export function mat_forward_textured_gouraud(
     gl: WebGL2RenderingContext
-): Material<ForwardInstancedLayout> {
+): Material<TexturedShadedLayout & ForwardShadingLayout> {
     let program = link(gl, vertex, fragment);
     return {
         Mode: GL_TRIANGLES,
@@ -89,14 +104,18 @@ export function mat2_forward_instanced(
             World: gl.getUniformLocation(program, "world")!,
             Self: gl.getUniformLocation(program, "self")!,
 
-            Palette: gl.getUniformLocation(program, "palette")!,
+            DiffuseMap: gl.getUniformLocation(program, "diffuse_map")!,
+            DiffuseColor: gl.getUniformLocation(program, "diffuse_color")!,
+            SpecularColor: gl.getUniformLocation(program, "specular_color")!,
+            Shininess: gl.getUniformLocation(program, "shininess")!,
 
+            Eye: gl.getUniformLocation(program, "eye")!,
             LightPositions: gl.getUniformLocation(program, "light_positions")!,
             LightDetails: gl.getUniformLocation(program, "light_details")!,
 
             VertexPosition: gl.getAttribLocation(program, "attr_position")!,
+            VertexTexCoord: gl.getAttribLocation(program, "attr_texcoord")!,
             VertexNormal: gl.getAttribLocation(program, "attr_normal")!,
-            VertexOffset: gl.getAttribLocation(program, "attr_offset")!,
         },
     };
 }
