+++ materials/mat_forward_textured_unlit.ts
@@ -33,7 +33,7 @@ let fragment = `#version 300 es\n
     }
 `;
 
-export function mat2_forward_textured_unlit(
+export function mat_forward_textured_unlit(
     gl: WebGL2RenderingContext
 ): Material<TexturedUnlitLayout> {
     let program = link(gl, vertex, fragment);
